backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-cors-headers>=4.0
Pillow>=10.0
django-storages>=1.14
boto3>=1.34
minio>=7.2
requests>=2.31
